refactor(auth): login and register in wizard
